add page examples for initial setup
