+++ front/src/components/Pages/CreatePostPage.tsx
@@ -1,0 +1,11 @@
+const CreatePostPage = () => {
+    return (
+        <div className='min-h-screen w-full flex flex-col justify-center items-center bg-gray-100 p-4'>
+            <h2 className='text-4xl'>
+                this is the create a post page
+            </h2>
+        </div>
+    )
+}
+
+export default CreatePostPage
