+++ front/src/components/Pages/PostsPage.tsx
@@ -1,17 +1,14 @@
-import {Link} from 'react-router-dom';
 
-const HomePage = () => {
+const PostsPage = () => {
     return (
-        <div className='min-h-screen w-full flex flex-col justify-center items-center bg-gray-100 p-4'>
-            <h1 className='text-4xl font-bold mb-6'>
-                this is the homepage
-            </h1>
-            <div className='flex gap-4'>
-                <Link to='/posts'>View posts</Link>
-                <Link to='/create'>Create a post</Link>
+        <div>
+            <div className='min-h-screen w-full flex flex-col justify-center items-center bg-gray-100 p-4'>
+                <h2 className='text-4xl'>
+                    this is the page where you can see all the posts
+                </h2>
             </div>
         </div>
     )
 }
 
-export default HomePage;
+export default PostsPage;
